Include <iostream> where necessary.
git-svn-id: https://svn.dealii.org/trunk@5422 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/deal.II/subdomain_ids.cc
@@ -164,7 +164,7 @@ void test ()
 		  ExcNumberMismatch(std::count (selected_dofs.begin(),
 						selected_dofs.end(),
 						true),
-				    pow(cells_per_direction/2+1,dim)));
+				    (int)pow(cells_per_direction/2+1,dim)));
 	}
       deallog << "Check 4 (dim=" << dim << ") ok" << std::endl;      
     };
